backend/requirements.txt
fastapi
uvicorn
google-genai
google-cloud-vision
pydantic
python-dotenv
pytest
python-multipart
transformers
torch
pillow
